requirements.txt
numpy==1.22.4
pandas==1.4.2
Pillow==9.2.0
plotly==5.8.0
streamlit==1.12.2
